+++ src/Monify.Tests/Strategies/ConvertFromStrategyTests/WhenGenerateIsCalled.cs
@@ -10,7 +10,7 @@ public sealed class WhenGenerateIsCalled
     {
         // Arrange
         Subject subject = TestSubject.Create();
-        subject.Conversions = [new Conversion { HasConversionFrom = true, Type = "int" }];
+        subject.Encapsulated = [new Encapsulated { HasConversionFrom = true, Type = "int" }];
         var strategy = new ConvertFromStrategy();
 
         // Act
@@ -40,7 +40,7 @@ public sealed class WhenGenerateIsCalled
     {
         // Arrange
         Subject subject = TestSubject.Create();
-        subject.Conversions = [new Conversion { Type = "int" }, new Conversion { Type = "string" }];
+        subject.Encapsulated = [new Encapsulated { Type = "int" }, new Encapsulated { Type = "string" }];
         var strategy = new ConvertFromStrategy();
 
         // Act
@@ -49,7 +49,7 @@ public sealed class WhenGenerateIsCalled
         // Assert
         sources.Length.ShouldBe(2);
         sources[0].Hint.ShouldBe("ConvertFrom");
-        sources[1].Hint.ShouldBe("ConvertFrom.Nested.0");
+        sources[1].Hint.ShouldBe("ConvertFrom.Passthrough.Level01");
         sources[1].Code.ShouldContain("implicit operator string(");
     }
 }