+++ src/Monify.Tests/TestSubject.cs
@@ -8,13 +8,12 @@ internal static class TestSubject
     {
         return new()
         {
-            Conversions = [new Conversion { Type = "int" }],
+            Encapsulated = [new Encapsulated { Type = "int" }],
             Declaration = "class",
             Name = "Sample",
             Namespace = string.Empty,
             Nesting = [],
             Qualification = "Sample",
-            Value = "int",
         };
     }
 }