+++ src/Monify/Strategies/InequalityStrategy.cs
@@ -1,41 +1,49 @@
-﻿namespace Monify.Strategies;
+namespace Monify.Strategies;
 
+using System.Collections.Generic;
 using Monify.Model;
 
 /// <summary>
 /// Generates the source needed to support the inequality operator.
 /// </summary>
-internal sealed class InequalityStrategy
+internal sealed partial class InequalityStrategy
     : IStrategy
 {
-    private readonly Predicate<Subject> _condition;
-    private readonly string _name;
-    private readonly Func<Subject, string> _type;
-
-    /// <summary>
-    /// Creates a new instance of the <see cref="InequalityStrategy"/>.
-    /// </summary>
-    /// <param name="condition">The condition for which, when <see langword="true" />, will result in code generation.</param>
-    /// <param name="name">The name of the inequality operator, used as part of the hint name for the generated code.</param>
-    /// <param name="type">The qualification for the type which serves as the subject for comparison.</param>
-    public InequalityStrategy(Predicate<Subject> condition, string name, Func<Subject, string> type)
-    {
-        _condition = condition;
-        _name = name;
-        _type = type;
-    }
+    private const int IndexForEncapsulatedValue = 0;
 
     /// <inheritdoc/>
     public IEnumerable<Source> Generate(Subject subject)
     {
-        if (!_condition(subject))
+        foreach (Operation operation in GetOperations(subject))
         {
-            yield break;
+            if (operation.HasOperator)
+            {
+                continue;
+            }
+
+            yield return new Source(CreateInequality(subject, operation.Type), operation.Hint);
         }
+    }
 
-        string type = _type(subject);
+    private static IEnumerable<Operation> GetOperations(Subject subject)
+    {
+        yield return new Operation(subject.HasInequalityOperator, "Inequality.Self", subject.Qualification);
 
-        string code = $$"""
+        for (int index = 0; index < subject.Encapsulated.Length; index++)
+        {
+            Encapsulated conversion = subject.Encapsulated[index];
+
+            string hint = index == IndexForEncapsulatedValue
+                ? "Inequality.Value"
+                : $"Inequality.Passthrough.Level{index:D2}";
+
+            yield return new Operation(conversion.HasInequalityOperator, hint, conversion.Type);
+        }
+    }
+
+    private static string CreateInequality(Subject subject, string type)
+    {
+        return $$"""
             {{subject.Declaration}} {{subject.Qualification}}
             {
                 public static bool operator !=({{subject.Qualification}} left, {{type}} right)
@@ -44,7 +52,5 @@ internal sealed class InequalityStrategy
                 }
             }
             """;
-
-        yield return new Source(code, $"Inequality.{_name}");
     }
 }